@@ -2,7 +2,7 @@ package com.szg_tech.heartcheck.activities.authentication
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.Log
 import android.view.animation.Animation
 import android.view.animation.AnimationUtils
