@@ -1,7 +1,7 @@
 package com.szg_tech.heartcheck.core.views.cell;
 
 import android.content.Context;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 import android.util.AttributeSet;
 import android.widget.ImageView;
 
