@@ -2,7 +2,7 @@ package com.szg_tech.heartcheck.core.views.cell;
 
 import android.content.Context;
 import android.graphics.Color;
-import android.support.v4.content.ContextCompat;
+import androidx.core.content.ContextCompat;
 import android.util.AttributeSet;
 import android.view.View;
 import android.widget.RadioButton;
