@@ -2,8 +2,8 @@ package com.szg_tech.heartcheck.core.views.modal;
 
 import android.content.Context;
 import android.content.res.Resources;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.TextView;
