@@ -4,15 +4,14 @@ import android.app.Activity;
 import android.content.res.Resources;
 import android.graphics.drawable.ColorDrawable;
 import android.os.Bundle;
-import android.support.design.widget.Snackbar;
-import android.support.design.widget.SnackbarContentLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.util.TypedValue;
 import android.view.Gravity;
@@ -23,6 +22,7 @@ import android.widget.FrameLayout;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
+import com.google.android.material.snackbar.SnackbarContentLayout;
 import com.szg_tech.heartcheck.R;
 import com.szg_tech.heartcheck.activities.evaluation.EvaluationActivity;
 import com.szg_tech.heartcheck.core.AbstractPresenter;
@@ -503,7 +503,7 @@ class EvaluationListPresenterImpl extends AbstractPresenter<EvaluationListView> 
         ((SnackbarContentLayout) snackbarView.getChildAt(0)).setGravity(Gravity.CENTER);
         ((FrameLayout.LayoutParams) snackbarView.getChildAt(0).getLayoutParams()).gravity = Gravity.CENTER;
         snackbarView.getLayoutParams().height = getView().getBottomButton().getHeight();
-        TextView snackBarTextView = snackbarView.findViewById(android.support.design.R.id.snackbar_text);
+        TextView snackBarTextView = snackbarView.findViewById(R.id.snackbar_text);
         snackBarTextView.setTextAlignment(View.TEXT_ALIGNMENT_CENTER);
         snackBarTextView.setGravity(Gravity.CENTER);
         ((LinearLayout.LayoutParams) snackBarTextView.getLayoutParams()).gravity = Gravity.CENTER;
