@@ -1,6 +1,6 @@
 package com.szg_tech.heartcheck.fragments.home;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.szg_tech.heartcheck.core.MVPView;
 
