@@ -1,8 +1,8 @@
 package com.szg_tech.heartcheck.fragments.register;
 import android.app.Dialog;
 import android.os.Bundle;
-import android.support.v4.app.DialogFragment;
-import android.support.v7.app.AlertDialog;
+import androidx.fragment.app.DialogFragment;
+import androidx.appcompat.app.AlertDialog;
 import android.text.Html;
 
 import com.szg_tech.heartcheck.R;
