@@ -1,6 +1,6 @@
 package com.szg_tech.heartcheck.fragments.saved_evaluation_list;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.szg_tech.heartcheck.core.MVPView;
